Rename WeakLibReader.hpp to AxisTypes.hpp
Clarifies that AxisTypes.hpp contains core types (Axis, InterpConfig,
AxisScale, OutOfRangePolicy) while LogInterpolate.hpp is the primary
user-facing interpolation API.
