+++ AGENTS.md
@@ -9,7 +9,7 @@ Translate WeakLib’s EOS & opacity **interpolators** from Fortran into **GPU‑
 ```
 WeakLibReader/
   src/                        # Core headers (public API + device helpers)
-    WeakLibReader.hpp         # Axis metadata, layout, out-of-range policy, internal interpolation helpers
+    AxisTypes.hpp             # Core types: Axis, InterpConfig, AxisScale, OutOfRangePolicy
     IndexDelta.hpp            # Linear/log10 indexing helpers
     InterpBasis.hpp           # Linear/bi-/tri-/tetra-/penta-linear basis routines
     InterpLogTable.hpp        # Log-space point kernels and aligned slices
@@ -117,7 +117,7 @@ ctest --test-dir build --output-on-failure
 
 - **Layout & Strides:** `WeakLibReader/src/Layout.hpp` provides `Layout` and `MakeLayout` helpers that precompute row-major strides and sub-slice utilities (`SliceLeading`), keeping device functions inline and `noexcept`.
 - **Index Lookup:** `WeakLibReader/src/IndexDelta.hpp` implements `IndexAndDeltaLin/Log10`, returning clamped cell indices and interpolation fractions plus an out-of-range flag to honor the configured policy.
-- **Interpolation Kernels:** `WeakLibReader/src/InterpBasis.hpp` supplies linear through penta-linear blending and partial derivatives. `WeakLibReader/src/WeakLibReader.hpp` provides axis metadata (`Axis`, `MakeAxis`), out-of-range policy (`InterpConfig`), and internal helpers used by higher-level APIs.
+- **Interpolation Kernels:** `WeakLibReader/src/InterpBasis.hpp` supplies linear through penta-linear blending and partial derivatives. `WeakLibReader/src/AxisTypes.hpp` provides axis metadata (`Axis`, `MakeAxis`), out-of-range policy (`InterpConfig`), and internal helpers used by higher-level APIs.
 - **Log-Wrapped APIs:** `WeakLibReader/src/InterpLogTable.hpp` and `WeakLibReader/src/LogInterpolate.hpp` layer pow10/offset handling, symmetric plane helpers, and derivative evaluators that replicate the Fortran log-table entry points.
 - **HDF5 Loader:** `WeakLibReader/src/Hdf5Loader.hpp` reads axis metadata + value datasets, validates monotonicity, and materializes tables into `amrex::TableData<double,4>` while preserving axis storage for interpolation.
 - **Build & Tests:** `CMakeLists.txt` exposes the headers as an INTERFACE target with AMReX/OpenMP/HDF5 includes; `test/test_log_interpolate.cpp` and `test/test_hdf5_loader.cpp` cover interpolation kernels, policies, symmetry helpers, weighted sums, derivatives, and HDF5 round-trips via the bundled Catch shim.
