+++ WeakLibReader/src/AxisTypes.hpp
